using uint8_t

+++ src/shared/CarduinoNode/MediaControlCarduinoNode/MediaControlCarduinoNode.cpp
@@ -12,7 +12,7 @@
  * 	to send string messages to radio use ids: on adroid app store a json file mapping id to message
  */
 
-MediaControlCarduinoNode::MediaControlCarduinoNode(uint8_t id, uint8_t clk, uint8_t dt, uint8_t sw, int cs, int interruptPin, uint8_t digiPotCs, uint8_t digiPotUd, uint8_t digiPotInc, uint8_t buzzer, const char *ssid, const char *password) : CarduinoNode(id, cs, interruptPin, ssid,  password, false, true, true) {
+MediaControlCarduinoNode::MediaControlCarduinoNode(uint8_t id, uint8_t clk, uint8_t dt, uint8_t sw, uint8_t cs, uint8_t interruptPin, uint8_t digiPotCs, uint8_t digiPotUd, uint8_t digiPotInc, uint8_t buzzer, const char *ssid, const char *password) : CarduinoNode(id, cs, interruptPin, ssid,  password, false, true, true) {
     this->restoreSettings();
 
 	this->canExecutor->addExecutor(new MediaControlCanEvent());
